requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain-google-genai==1.0.10
langchain==0.1.20
python-dotenv==1.0.0
browser-use==0.1.0
playwright==1.40.0
asyncio